+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # vNext
 
-* Enhancement: Set transaction name on events and transactions sent using Spring integration (#1067) 
+* Fixes and Tests: Session serialization and deserialization
+
+# 4.0.0-alpha.2
+
+* Feat: Add basic support for attachments (#1082)
+* Fix: Remove method reference in SentryEnvelopeItem (#1091)
+* Enhancement: Set transaction name on events and transactions sent using Spring integration (#1067)
 * Fix: Set current thread only if there are no exceptions
 * Enhancement: Set global tags on SentryOptions and load them from external configuration (#1066)
 * Ref: Refactor resolving SpanContext for Throwable (#1068)
@@ -10,6 +16,10 @@
 * Enhancement: Resolve in-app-includes and in-app-excludes parameters from the external configuration
 * Enhancement: Make InAppIncludesResolver public (#1084)
 * Ref: Change "op" to "operation" in @SentrySpan and @SentryTransaction
+* Fix: SentryOptions creates GsonSerializer by default
+* Enhancement: Add the ability to register multiple OptionsConfiguration beans (#1093)
+* Fix: Append DebugImage list if event already has it
+* Fix: Sort breadcrumbs by Date if there are breadcrumbs already in the event
 
 # 4.0.0-alpha.1